Move timer to ctrl instead model
